% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_ccsp_by_county.R
\name{fit_ccsp_by_county}
\alias{fit_ccsp_by_county}
\title{Load the FIT and CCSP data by county}
\usage{
fit_ccsp_by_county(county)
}
\description{
Load the FIT and CCSP data by county
}
\examples{

d <- fit_ccsp_by_county()

head(d)

summary(d)

}
