% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fit_and_ccsp_activity.R
\name{map_fit_and_ccsp_activity}
\alias{map_fit_and_ccsp_activity}
\title{map of FIT and CCSP activity}
\usage{
map_fit_and_ccsp_activity()
}
\description{
map of FIT and CCSP activity
}
\examples{

map_fit_and_ccsp_activity()

}
